+++ Gemfile
@@ -72,3 +72,5 @@ gem 'jquery-rails'
 gem 'font-awesome-rails'
 
 gem 'pry-rails'
+
+gem "mutool"
