+++ app/controllers/documents_controller.rb
@@ -1,30 +1,9 @@
 class DocumentsController < ApplicationController
   def index 
-    @user_name = current_user.name
-    
-  end
+    @folders  = Folder.all
+    @folder = Folder.find(params[:folder_id])
+    @document = @folder.documents
 
-  def new
-    @document = DocumentFolder.new
-  end
-
-  def create
-    @document = DocumentFolder.new(document_params)
-    @folder = Folder.new
-
-    if @document.valid?
-      @document.save
-      binding.pry
-      return redirect_to root_path   
-    else
-      render "new"
-    end
-  end
-
-
-  private
-  def document_params
-    params.require(:document_folder).permit(:document_name, :folder_name, :image).merge(user_id: current_user.id)
   end
   
 end
