+++ app/helpers/folders_helper.rb
@@ -1,0 +1,2 @@
+module FoldersHelper
+end
