+++ app/views/devise/sessions/new.html.erb
@@ -8,7 +8,6 @@
             <h5 class="card-title text-center">Sign In</h5>
             <%= form_for(resource, as: resource_name, url: session_path(resource_name), class: "form-signin") do |f| %>
               <%= render "devise/shared/error_messages", resource: resource %>
-              <%= flash[:alert] %>
               <div class="form-label-group">
                 <%= f.email_field :email, id: "inputEmail", class: "form-control", placeholder: "Email address", autofocus: true, autocomplete: "email"%>
                 <%= f.label :email, for: "inputEmail"%>
