+++ app/views/folders/index.html.erb
@@ -7,7 +7,7 @@
       <%= render "documents/folders" %>
     </div>
     <div class="col">
-      <%= render "documents/document" %>
+
     </div>
   </div>
 </div>
