+++ app/views/folders/new.html.erb
@@ -6,8 +6,20 @@
       </div>
          <%= link_to "Log out", destroy_user_session_path, class: "btn btn-lg btn-facebook", span: "sr-only", method: :delete %>
     </div>
-    <%= form_with model: @document, url: documents_path do |f| %>
-    <%= render 'layouts/error_messages', document: @document %>
+    <%= form_with model: @document, url: folders_path do |f| %>
+    
+    <% if @document.errors.any? %>
+    <div class="alert error-alert">
+      <ul>
+        <% @document.errors.full_messages.each do |message| %>
+          <li class="error-message">
+            <%= message %>
+          </li>
+        <% end %>
+      </ul>
+    </div>
+    <% end %>
+    
       <div class="form-row">
         <div class="form-group col-sm-4">
           <%= f.label :folder_name, "Folder name" %>
@@ -23,7 +35,7 @@
         </div>
         <div class="form-group col-4">
           <div class="form-group" style="margin-top: 21px;">
-            <%= f.file_field :image %>
+            <%= f.file_field :images,  multiple: true %>
             <%= f.submit '送信する', class: "form-submit", accept:".pdf" %>
             <!-- <form method="post" enctype="multipart/form-data">
               <input type="file" accept=".pdf" name="avatar">
