--- app/views/layouts/_side_bar.html.erb
@@ -1,18 +1,0 @@
-<div class="folders  overflow-hidden" style="margin-left: 10px;">
-  <% current_user.folders.each do |folder| %>
-    <div class="folder border" style="padding: 10px; width: 200px; margin-top: 10px;">
-      <a href="#"><%= folder.folder_name %></a>
-    </div>
-  <% end %>
-</div>
-
-
-
-<!-- <div class="folders">
-      <p class="text-center">folder</p>
-    <div class="folder">
-      <div class="folder-name">
-        
-      </div>
-    </div>
-</div> -->
