+++ db/migrate/20200923104731_rename_name_column_to_folders.rb
@@ -1,6 +1,5 @@
 class RenameNameColumnToFolders < ActiveRecord::Migration[6.0]
   def change
     rename_column :folders, :name, :folder_name
-    add_index     :folders, :folder_name, unique: true
   end
 end
