+++ spec/requests/folders_request_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "Folders", type: :request do
+
+end
